Fix versioning and add module export feature

+++ sfcompiler.js
@@ -1,7 +1,6 @@
 module.exports = function(obj, gulp){
 // Load variables
 var path = obj.path;
-var compilePath = obj.compilePath;
 var includeSourceMap = obj.includeSourceMap;
 
 // Load dependency
@@ -23,164 +22,225 @@ var sass = null;
 
 var compiling = false;
 
+function init(){
+	console.log("[Preparing] .js handler");
+	prepareJS();
+
+	console.log("[Preparing] .scss handler");
+	prepareSCSS();
+
+	console.log("[Preparing] .html handler");
+	prepareHTML();
+}
+
 // === Javascript Recipe ===
 // 
-gulp.task('watch-js', function(done){
-	var js = gulp.series('js');
-	var last = 0;
-	gulp.watch(compilePath.js).on('change', function(file, stats){
-		if(last === stats.ctimeMs)
-			return;
+function prepareJS(){
+	watchPath('js', function(name, obj){
+		var last = 0;
 
-		last = stats.ctimeMs;
-		if(SFLang.scan(file, stats))
-			return;
+		name = 'js-'+name;
+		gulp.task(name, jsTask(obj));
 
-		js();
+		var call = gulp.series(name);
+		if(compiling === false)
+			gulp.watch(obj.js.combine).on('change', function(file, stats){
+				if(last === stats.ctimeMs)
+					return;
+
+				last = stats.ctimeMs;
+				if(SFLang.scan(file, stats))
+					return;
+
+				call();
+			});
+
+		else call();
 	});
-	done();
-});
-gulp.task('js', function(){
-	removeOldMap(path.js.folder, path.js.file.replace('.js', ''), '.js');
-	var temp = gulp.src(compilePath.js);
+}
+function jsTask(path){
+	return function(){
+		var startTime = Date.now();
+		removeOldMap(path.js.folder, path.js.file.replace('.js', ''), '.js');
+		var temp = gulp.src(path.js.combine);
 
-	if(includeSourceMap)
-		temp = temp.pipe(sourcemaps.init());
+		if(includeSourceMap)
+			temp = temp.pipe(sourcemaps.init());
 
-	temp = temp.pipe(concat(path.js.file)).pipe(SFLang.jsPipe());
+		temp = temp.pipe(concat(path.js.file)).pipe(SFLang.jsPipe());
 
-	if(compiling){
-		if(!uglify) uglify = require('gulp-uglify-es').default;
-		if(!babel) babel = require('gulp-babel');
+		if(compiling){
+			if(!uglify) uglify = require('gulp-uglify-es').default;
+			if(!babel) babel = require('gulp-babel');
 
-		temp = temp.pipe(babel({
-			babelrc: false,
-			presets: [
-				["@babel/preset-env", {
-					targets: {
-						ie: "11"
-					},
-					modules: false,
-					loose: true
-				}]
-			]
-		})).on('error', swallowError).pipe(uglify({mangle: {toplevel: true}})).on('error', swallowError);
+			temp = temp.pipe(babel({
+				babelrc: false,
+				presets: [
+					["@babel/preset-env", {
+						targets: {
+							ie: "11"
+						},
+						modules: false,
+						loose: true
+					}]
+				]
+			})).on('error', swallowError).pipe(uglify({mangle: {toplevel: true}})).on('error', swallowError);
+		}
+
+		if(path.js.header)
+			temp = temp.pipe(header(path.js.header+"\n"));
+
+		if(includeSourceMap)
+			temp = temp.pipe(sourcemaps.write('.', {
+				mapFile: function(mapFilePath) {
+					return mapFilePath.replace('js.map', startTime+'.js.map');
+				}
+			}))
+
+		temp = temp.pipe(gulp.dest(path.js.folder)).on('end', function(){
+				if(notifier)
+					notifier.notify({
+						title: 'Gulp Compilation',
+						message: 'JavaScript was finished!'
+					});
+
+				if(browserSync)
+					browserSync.reload(path.js.folder+path.js.file);
+			});
+
+		versioning(path.versioning, path.js.folder.replace(path.stripURL || '#$%!.', '')+path.js.file+'?', startTime);
+		return temp;
 	}
-
-	if(includeSourceMap)
-		temp = temp.pipe(sourcemaps.write('.', {
-			mapFile: function(mapFilePath) {
-				return mapFilePath.replace('js.map', Date.now()+'.js.map');
-			}
-		}))
-
-	temp = temp.pipe(gulp.dest(path.js.folder)).on('end', function(){
-			if(notifier)
-				notifier.notify({
-					title: 'Gulp Compilation',
-					message: 'JavaScript was finished!'
-				});
-
-			if(browserSync)
-				browserSync.reload(path.js.folder+path.js.file);
-		});
-
-	versioning(path.versioning, path.js.folder.replace('public/', '')+path.js.file+'?');
-	return temp;
-});
+}
 
 // === SCSS Recipe ===
 // 
-gulp.task('watch-css', function(done){
-	gulp.watch(compilePath.css, gulp.series('scss'));
-	done();
-});
-gulp.task('scss', function(){
+function prepareSCSS(){
+	watchPath('scss', function(name, obj){
+		var last = 0;
+
+		name = 'scss-'+name;
+		gulp.task(name, scssTask(obj));
+
+		var call = gulp.series(name);
+		if(compiling === false)
+			gulp.watch(obj.scss.combine).on('change', function(file, stats){
+				if(last === stats.ctimeMs)
+					return;
+
+				last = stats.ctimeMs;
+				call();
+			});
+
+		else call();
+	});
+}
+function scssTask(path){
 	if(!sass) sass = require('gulp-sass');
 
-	removeOldMap(path.css.folder, path.css.file.replace('.css', ''), '.css');
-	var temp = gulp.src(compilePath.css);
+	return function(){
+		var startTime = Date.now();
+		removeOldMap(path.scss.folder, path.scss.file.replace('.css', ''), '.css');
+		var temp = gulp.src(path.scss.combine);
 
-	if(includeSourceMap)
-		temp = temp.pipe(sourcemaps.init());
+		if(includeSourceMap)
+			temp = temp.pipe(sourcemaps.init());
 
-	temp = temp.pipe(sass()).on('error', swallowError);
+		temp = temp.pipe(sass()).on('error', swallowError);
 
-	if(compiling){
-		if(!csso) csso = require('gulp-csso');
-		if(!autoprefixer) autoprefixer = require('gulp-autoprefixer');
+		if(compiling){
+			if(!csso) csso = require('gulp-csso');
+			if(!autoprefixer) autoprefixer = require('gulp-autoprefixer');
 
-		temp = temp.pipe(autoprefixer()).pipe(csso());
+			temp = temp.pipe(autoprefixer()).pipe(csso());
+		}
+
+		temp = temp.pipe(concat(path.scss.file));
+
+		if(path.scss.header)
+			temp = temp.pipe(header(path.scss.header+"\n"));
+
+		if(includeSourceMap)
+			temp = temp.pipe(sourcemaps.write('.', {
+				mapFile: function(mapFilePath) {
+					return mapFilePath.replace('css.map', startTime+'.css.map');
+				}
+			}));
+
+		temp = temp.pipe(gulp.dest(path.scss.folder)).on('end', function(){
+				if(notifier)
+					notifier.notify({
+						title: 'Gulp Compilation',
+						message: 'SCSS was finished!'
+					});
+
+				if(browserSync){
+					browserSync.reload(path.scss.folder+path.scss.file);
+					browserSync.notify("SCSS Reloaded");
+				}
+			});
+
+		versioning(path.versioning, path.scss.folder.replace(path.stripURL || '#$%!.', '')+path.scss.file+'?', startTime);
+		return temp;
 	}
-
-	temp = temp.pipe(concat(path.css.file));
-
-	if(includeSourceMap)
-		temp = temp.pipe(sourcemaps.write('.', {
-			mapFile: function(mapFilePath) {
-				return mapFilePath.replace('css.map', Date.now()+'.css.map');
-			}
-		}));
-
-	temp = temp.pipe(gulp.dest(path.css.folder)).on('end', function(){
-			if(notifier)
-				notifier.notify({
-					title: 'Gulp Compilation',
-					message: 'SCSS was finished!'
-				});
-
-			if(browserSync){
-				browserSync.reload(path.css.folder+path.css.file);
-				browserSync.notify("SCSS Reloaded");
-			}
-		});
-
-	versioning(path.versioning, path.css.folder.replace('public/', '')+path.css.file+'?');
-	return temp;
-});
+}
 
 // === HTML Recipe ===
 // 
-gulp.task('watch-html', function(done){
-	var last = 0
+function prepareHTML(){
+	watchPath('html', function(name, obj){
+		var last = 0;
 
-	// Combine all html into one js file
-	var html = gulp.series('html');
-	gulp.watch(compilePath.html).on('change', function(file, stats){
-		if(last === stats.ctimeMs)
-			return;
+		name = 'html-'+name;
+		gulp.task(name, htmlTask(obj));
 
-		last = stats.ctimeMs;
-		SFLang.scan(file, stats);
+		var call = gulp.series(name);
+		if(compiling === false){
+			if(obj.static !== void 0){
+				gulp.watch(obj.html.combine).on('change', function(file, stats){
+					if(last === stats.ctimeMs)
+						return;
 
-		html();
+					last = stats.ctimeMs;
+					SFLang.scan(file, stats);
+				});
+
+				obj.combine = excludeSource(obj.combine, obj.static);
+			}
+
+			gulp.watch(obj.html.combine).on('change', function(file, stats){
+				if(last === stats.ctimeMs)
+					return;
+
+				last = stats.ctimeMs;
+				SFLang.scan(file, stats);
+
+				call();
+			});
+		}
+
+		else call();
 	});
+}
+function htmlTask(path){
+	return function(){
+		var startTime = Date.now();
+		versioning(path.versioning, path.html.folder.replace(path.stripURL || '#$%!.', '')+path.html.file+'?', startTime);
 
-	// Watch changes on PHP/HTML
-	if(compilePath.template){
-		gulp.watch(compilePath.template).on('change', function(file, stats){
-			if(last === stats.ctimeMs)
-				return;
-
-			last = stats.ctimeMs;
-			SFLang.scan(file, stats);
-		});
+		return gulp.src(path.html.combine)
+			.pipe(htmlToJs({global:'window.templates', concat:path.html.file, prefix:path.html.prefix}))
+			.pipe(header((path.html.header+"\n" || '') + "\nif(window.templates === void 0)"))
+			.pipe(gulp.dest(path.html.folder));
 	}
-	done();
-});
-gulp.task('html', function(){
-	versioning(path.versioning, path.html.folder.replace('public/', '')+path.html.file+'?');
-
-	return gulp.src(compilePath.html)
-		.pipe(htmlToJs({global:'window.templates', concat:path.html.file}))
-		.pipe(gulp.dest(path.html.folder));
-});
+}
 
 // === Other ===
 // 
 gulp.task('browser-sync', function(){
 	if(!obj.browserSync)
 		return;
+
+	init();
 
 	notifier = require('node-notifier');
 	browserSync = require('browser-sync');
@@ -189,22 +249,15 @@ gulp.task('browser-sync', function(){
 });
 
 // To be executed on Development computer
-gulp.task('default', gulp.parallel(['browser-sync', 'watch-css', 'watch-js', 'watch-html']));
+gulp.task('default', gulp.series('browser-sync'));
 
 // === Compiling Recipe ===
-// 
-gulp.task('compiling', function(done){
+// To be executed on Continuous Delivery
+gulp.task('compile', function(done){
 	compiling = true;
-	includeSourceMap = false;
+	init();
 	done();
 });
-
-// To be executed on Continuous Delivery
-gulp.task('compile', gulp.series('compiling', gulp.parallel(['js', 'scss', 'html'])));
-
-gulp.task('compilejs', gulp.series(['compiling', 'js']));
-gulp.task('compilecss', gulp.series(['compiling', 'scss']));
-gulp.task('extractlang', gulp.series(['compiling', 'scss']));
 
 
 // === No need to edit below ===
@@ -213,20 +266,79 @@ function swallowError(error){
 	this.emit('end');
 }
 
-function versioning(target, prefixStart){
+function versioning(target, prefixStart, timestamp){
 	var regex = prefixStart.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
 
 	var data = fs.readFileSync(target, 'utf8');
-	fs.writeFileSync(target, data.replace(RegExp(regex + '[0-9]+', 'g'), prefixStart+Date.now()), 'utf8');
+	fs.writeFileSync(target, data.replace(RegExp(regex + '[0-9a-z]+', 'g'), prefixStart+timestamp), 'utf8');
 }
 
 function removeOldMap(path, filename, format){
 	fs.readdir(path, function(err, files){
+		if(files === void 0)
+			return;
+
 		for (var i = 0, len = files.length; i < len; i++) {
 			if(files[i].indexOf(filename) === 0 && files[i].indexOf(format+'.map') !== -1)
 				fs.unlinkSync(path+files[i]);
 		}
 	});
+}
+
+function excludeSource(old, news){
+	if(old.constructor !== Array)
+		old = [old];
+
+	if(news.constructor === Array){
+		for (var i = 0; i < news.length; i++) {
+			old.push('!'+news[i]);
+		}
+	}
+	else old.push(news);
+
+	return old;
+}
+
+function splitFolderPath(fullPath){
+	fullPath = fullPath.replace(/\\/g, '/').split('/');
+	var file = fullPath.pop();
+	var folder = fullPath.join('/');
+
+	if(folder.length !== 0)
+		folder += '/';
+
+	return [file, folder];
+}
+
+function watchPath(which, watch){
+	var default_ = path.default;
+	delete path.default;
+
+	var list = Object.keys(path);
+	for (var i = 0; i < list.length; i++) {
+		var temp = path[list[i]];
+
+		// Check if default was exist
+		if(default_ && default_[which])
+			default_[which].combine = excludeSource(default_[which].combine, temp[which].combine);
+
+		// Separate file name and folder path
+		temp[which].file = splitFolderPath(temp[which].file);
+		temp[which].folder = temp[which].file.pop();
+		temp[which].file = temp[which].file[0];
+
+		watch(list[i], temp);
+	}
+
+	if(default_){
+		// Separate file name and folder path
+		default_[which].file = splitFolderPath(default_[which].file);
+		default_[which].folder = default_[which].file.pop();
+		default_[which].file = default_[which].file[0];
+
+		path.default = default_;
+		watch('default', default_);
+	}
 }
 
 };
